Move hashing to a autogluon.timeseries.utils.hashing
