+++ timeseries/tests/unittests/test_ts_dataset.py
@@ -698,44 +698,3 @@ SAMPLE_DATAFRAME_WITH_MIXED_INDEX = pd.DataFrame(
 def test_when_item_id_index_has_mixed_dtype_then_value_error_is_raied(input_df):
     with pytest.raises(ValueError, match="must be of integer or string dtype"):
         TimeSeriesDataFrame(input_df)
-
-
-def test_given_dfs_are_identical_when_hashing_then_hash_is_identical():
-    df = SAMPLE_TS_DATAFRAME
-    df_copy = SAMPLE_TS_DATAFRAME.copy()
-    assert df.values.base is not df_copy.values.base
-    assert (df.hash_each_item().values == df_copy.hash_each_item().values).all()
-
-
-def test_given_item_id_is_different_when_hashing_then_hash_is_identical():
-    df1 = _build_ts_dataframe(item_ids=ITEM_IDS, datetime_index=DATETIME_INDEX, target=TARGETS)
-    df2 = _build_ts_dataframe(item_ids=["A", "B", "C"], datetime_index=DATETIME_INDEX, target=TARGETS)
-    assert (df1.hash_each_item().values == df2.hash_each_item().values).all()
-
-
-def test_given_timestamp_is_different_when_hashing_then_hash_is_different():
-    df1 = _build_ts_dataframe(item_ids=ITEM_IDS, datetime_index=DATETIME_INDEX, target=TARGETS)
-    df2 = _build_ts_dataframe(
-        item_ids=ITEM_IDS,
-        datetime_index=tuple(pd.date_range(pd.Timestamp("01-01-2020", freq="D"), periods=3)),
-        target=TARGETS,
-    )
-    assert (df1.hash_each_item().values != df2.hash_each_item().values).all()
-
-
-def test_given_values_are_different_when_hashing_then_hash_is_different():
-    df1 = _build_ts_dataframe(item_ids=ITEM_IDS, datetime_index=DATETIME_INDEX, target=TARGETS)
-    df2 = _build_ts_dataframe(item_ids=ITEM_IDS, datetime_index=DATETIME_INDEX, target=TARGETS * 2)
-    assert (df1.hash_each_item().values != df2.hash_each_item().values).all()
-
-
-def test_when_hashing_then_item_id_order_is_preserved():
-    df = _build_ts_dataframe(item_ids=["C", "1", "A"], datetime_index=DATETIME_INDEX, target=TARGETS)
-    assert df.hash_each_item().index.equals(df._item_index)
-
-
-def test_when_items_are_permuted_then_hash_values_are_permuted():
-    new_order = [2, 0, 1]
-    df1 = _build_ts_dataframe(item_ids=ITEM_IDS, datetime_index=DATETIME_INDEX, target=TARGETS)
-    df2 = df1.loc[new_order]
-    assert (df1.hash_each_item().loc[new_order].values == df2.hash_each_item().values).all()
